一些零散bug

+++ Assets/Scripts/ObtainableObjectData_and_Backpack/ObtainableObjectData/Editor/Distribute_Edi.cs
@@ -13,14 +13,14 @@ public class Distribute_Edi : Editor
 
         if (GUILayout.Button("Test_Dis_Collection_Random"))
         {
-            Collection_Data Data = Distributor.DistributeRandomCollectionbyLevel(Distributor.CollectionLevel);
+            Collection_Data Data = Distributor.DistributeRandomCollectionbyLevel((int)Distributor.CollectionLevel);
             Distributor.DistributeCollection(Distributor.gameObject.transform.position,GameObject.FindGameObjectWithTag("Player").transform.position,Data);
             Debug.Log(Data.Name);
         }
 
         if (GUILayout.Button("Test_Dis_Prop_Random"))
         {
-            Prop_Data Data = Distributor.DistributeRandomPropbyLevel(Distributor.PropLevel);
+            Prop_Data Data = Distributor.DistributeRandomPropbyLevel((int)Distributor.PropLevel);
             Distributor.DistributeProp(Distributor.gameObject.transform.position,GameObject.FindGameObjectWithTag("Player").transform.position, Data);
             Debug.Log(Data?.Name);
         }
